+++ qsprpred/data/data.py
@@ -172,7 +172,8 @@ class PandasDataSet(DataSet):
         """
         n_cpus = self.nJobs
         chunk_size = self.chunkSize
-        if n_cpus and n_cpus > 1:
+        if n_cpus and n_cpus > 1 and not (
+                hasattr(func, 'no_parallelization') and getattr(func, 'no_parallelization') is True):
             return self.papply(func, func_args, func_kwargs, axis, raw, result_type, subset, n_cpus, chunk_size)
         else:
             df_sub = self.df[subset if subset else self.df.columns]